All completed. Next plan Login.Register = RoleBased Actions = First warehouse the registration

+++ ITAsset-DDLA/ITAsset-DDLA/Helpers/Extentions/FileExtention.cs
@@ -41,6 +41,7 @@ public static class FileExtention
 
     public static void RemoveFile(string path)
     {
+        if (File.Exists(path))
         System.IO.File.Delete(path);
     }
